Split tests related to ROI items and core
